fixes bug with overflow
this basically resulted in a crap ton of NaN checking, there is also one
little hacky change that I had to make because for some reason calling
the `<` function was resulting in an integer overflow because of some
casting. I don't actually know what caused it, so I just treated the
symptom. oh well